NP-2919 Create base component for filters
